requirements.txt
numpy
faiss-cpu
pandas
pdfplumber
pymupdf
pytesseract
pdf2image
nltk
sentence-transformers
scikit-learn
fastapi
uvicorn
langchain
transformers
huggingface-hub
google-generativeai
openai
tqdm
python-dotenv
requests
PyYAML
